added class for as7263 spectral sensor with read and write functions

+++ obc/include/I2C.h
@@ -45,7 +45,7 @@ public:
 	I2CDevice(unsigned int bus, unsigned int device);
 	virtual int open();
 	virtual int write(unsigned char value);
-	virtual uint16_t readRegister(unsigned int registerAddress);
+	virtual uint8_t readRegister(unsigned int registerAddress);
 	virtual unsigned char* readRegisters(unsigned int number, unsigned int fromAddress=0);
 	virtual int writeRegister(unsigned int registerAddress, unsigned char value);
 	virtual void debugDumpRegisters(unsigned int number = 0xff);
